refactor all the states code into the states sub-package and import all the state classes into the states.__init__
