+++ src/pyvalem/_data_atomic_weights.txt
