+++ src/pyvalem/_data_isotope_masses.txt
