+++ src/pyvalem/_special_cases.py
