+++ src/pyvalem/states/atomic_term_symbol.py
@@ -6,8 +6,8 @@ an HTML representation of the term symbol, etc.
 
 import pyparsing as pp
 
-from .state import State, StateParseError
-from .utils import parse_fraction, float_to_fraction
+from pyvalem.states._base_state import State, StateParseError
+from pyvalem._utils import parse_fraction, float_to_fraction
 
 atom_L_symbols = "S P D F G H I K L M N O Q R T U V W X Y Z".split()
