+++ src/pyvalem/states/generic_excited_state.py
@@ -5,7 +5,7 @@ an atom, ion or molecule as '*', '**', '***', '****', '5*', etc.
 
 import pyparsing as pp
 
-from .state import State, StateParseError
+from pyvalem.states._base_state import State, StateParseError
 
 integer = pp.Word(pp.nums)
 atom_int = integer.setResultsName("int")
