+++ src/pyvalem/states/racah_symbol.py
@@ -6,7 +6,7 @@ labels, creating an HTML representation of the term symbol, etc.
 
 import pyparsing as pp
 
-from .state import State, StateParseError
+from pyvalem.states._base_state import State, StateParseError
 
 integer = pp.Word(pp.nums)
 atom_principal = integer.setResultsName("principal")
