+++ tests/test_atomic_configurations.py
@@ -4,7 +4,7 @@ Unit tests for the atomic_configurations module of PyValem
 
 import unittest
 
-from pyvalem.atomic_configuration import AtomicConfiguration, AtomicConfigurationError
+from pyvalem.states.atomic_configuration import AtomicConfiguration, AtomicConfigurationError
 
 
 class AtomicConfigurationTest(unittest.TestCase):
